pythonProject/requirements.txt
beautifulsoup4==4.11.1
Flask==2.0.3
nltk==3.7
pandas~=1.3.5
numpy~=1.21.5
Pillow~=9.0.1
matplotlib~=3.5.1
Werkzeug~=2.1.1
plotly~=5.7.0
charts~=0.4.6
scikit-learn
scipy==1.8.0
wordcloud==1.8.1